added crispy forms

+++ src/blog/views.py
@@ -1,3 +1,4 @@
+from django.http.response import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from .models import Post, Like
 from .forms import CommentForm, PostForm
@@ -50,6 +51,9 @@ def post_detail(request, slug):
 def post_update(request, slug):
     obj = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
+    if request.user.id != obj.author.id:
+        # return HttpResponse("You're not authorized!!")
+        return redirect("blog:list")
     if form.is_valid():
         form.save()
         return redirect("blog:list")
@@ -63,6 +67,9 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        # return HttpResponse("You're not authorized!!")
+        return redirect("blog:list")
     if request.method == "POST":
         obj.delete()
         return redirect("blog:list")
